chore(color): move isBrowser under utils
