+++ packages/color/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './isBrowser';
 export * from './supportColor';
